Everything except for hazards
It appears we are almost complete! We have not done any debugging yet, but we have a program that allows the robot to navigate through a maze and drop off cargo on the other side while mapping where the robot went. Still need to add functionality that allows the robot to report hazards in a separate file. Not sure if we should use a baseline value or report the highest sensor value (first one makes more logical sense in theory, second one makes more sense in practice).