Ready for testing
All of the things that we can code before testing appears to be complete. The code to make the hazards file is done. We still will need to write the conversion from what the sensors are reading to what we need to give the TAs, but other than that it appears that the robot is fully functional.